refactor: checkCardType내부 로직 배열로 분리
