+++ hooks/src/lib/useCardType.ts
@@ -1,9 +1,7 @@
 import { useState } from 'react';
-import {
-  CARD_BRAND_NUMBER_LENGTH,
-  CARD_CONFIG,
-} from './constants/cardDataValidation';
+import { CARD_BRAND_NUMBER_LENGTH } from './constants/cardDataValidation';
 import formatFunctions from './constants/formatFunction';
+import cardTypeCheck from './constants/cardTypeChecks';
 
 export type CardType = {
   name: keyof typeof CARD_BRAND_NUMBER_LENGTH;
@@ -11,34 +9,8 @@ export type CardType = {
 };
 
 const checkCardType = (cardNumber: string) => {
-  const cardBrandNumber = parseInt(cardNumber.substring(0, 2), 10);
-  const cardBrandNumberThree = parseInt(cardNumber.substring(0, 3), 10);
-  const cardBrandNumberFour = parseInt(cardNumber.substring(0, 4), 10);
-  const cardBrandNumberSix = parseInt(cardNumber.substring(0, 6), 10);
-
-  if (Math.floor(cardBrandNumber / 10) === CARD_CONFIG.VISA) return 'Visa';
-  if (
-    CARD_CONFIG.MASTER.START <= cardBrandNumber &&
-    cardBrandNumber <= CARD_CONFIG.MASTER.END
-  )
-    return 'MasterCard';
-
-  if (cardBrandNumber === 36) {
-    return 'Diners';
-  }
-
-  if (cardBrandNumber === 34 || cardBrandNumber === 37) {
-    return 'AMEX';
-  }
-
-  if (
-    (cardBrandNumberSix >= 622126 && cardBrandNumberSix <= 622925) ||
-    (cardBrandNumberThree >= 624 && cardBrandNumberThree <= 626) ||
-    (cardBrandNumberFour >= 6282 && cardBrandNumberFour <= 6288)
-  ) {
-    return 'UnionPay';
-  }
-  return 'Empty';
+  const cardType = cardTypeCheck.find((cardType) => cardType.check(cardNumber));
+  return cardType ? cardType.name : 'Empty';
 };
 
 const useCardType = () => {
@@ -48,7 +20,8 @@ const useCardType = () => {
   });
 
   const cardTypeHandler = (value: string) => {
-    const cardTypeName = checkCardType(value);
+    const cardTypeName: keyof typeof CARD_BRAND_NUMBER_LENGTH =
+      checkCardType(value);
     setCardType({
       name: cardTypeName,
       maxLength: CARD_BRAND_NUMBER_LENGTH[cardTypeName],
